requirements.txt
pandas~=1.2.3
matplotlib~=3.4.1
python-dateutil~=2.8.1
pymongo~=3.11.3
bson~=0.5.10
requests~=2.25.1
schedule~=1.1.0